+++ test/httpwg-tests.js
@@ -1,21 +1,73 @@
 const expect = require('chai').expect;
 const Parser = require('../dist/parser');
-const binaryTests = require('./httpwg-tests/binary.json');
-const listTests = require('./httpwg-tests/list.json');
-const numberTests = require('./httpwg-tests/list.json');
-const stringTests = require('./httpwg-tests/string.json');
 const base32Encode = require('base32-encode');
+const fs = require('fs');
 
-function addTest(test) {
+describe('HTTP-WG tests', () => {
+
+  const testGroups = [
+    'binary',
+    'boolean',
+    'number',
+    'string',
+    'token',
+
+    'item',
+
+    'list',
+    'listlist',
+    'dictionary',
+    'param-list',
+
+    'key-generated',
+    'large-generated',
+    'string-generated',
+    'token-generated',
+  ];
+
+  for(const testGroup of testGroups) {
+
+    describe(testGroup, () => {
+
+      makeTestGroup(testGroup);
+
+    });
+
+  }
+
+
+});
+
+function makeTestGroup(testGroup) {
+
+  const fileName = testGroup + '.json';
+  const blob = fs.readFileSync(__dirname + '/httpwg-tests/' + fileName);
+  const tests = JSON.parse(blob);
+
+  for(const test of tests) {
+    makeTest(test);
+  }
+
+}
+function makeTest(test) {
 
   const parser = new Parser(test.raw.join(','));
 
+  const skipped = [
+    'too long integer',
+    'negative too long integer',
+  ];
+
   it(test.name, function() {
-    if (test.name==='non-zero pad bits') {
-      // It's not worth the trouble to support this.
-      this.skip();
+
+    if (skipped.includes(test.name)) {
+      // Not yet supporting this.
+      // see: https://github.com/httpwg/structured-header-tests/issues/9
+      this.skip('Can\'t support this yet');
     }
+
     let hadError = false;
+    let caughtError;
     let result;
     try {
       switch(test.header_type) {
@@ -25,66 +77,84 @@ function addTest(test) {
         case 'list' :
           result = parser.parseList();
           break;
+        case 'list-list' :
+          result = parser.parseListList();
+          break;
+        case 'param-list' :
+          result = parser.parseParamList();
+          break;
+        case 'dictionary' :
+          result = parser.parseDictionary();
+          break;
+        default:
+          throw new Error('Unsupported header type: ' + test.header_type);
       }
     } catch (e) {
       hadError = true;
-      //console.log(e);
+      caughtError = e;
     }
 
-    if (test.expected === false) {
+    if (test.must_fail) {
       expect(hadError).to.equal(true);
     } else {
-      expect(hadError).to.equal(false);
 
-      if(result instanceof Buffer) {
-        result = base32Encode(result, 'RFC4648');
+      if (hadError) {
+        // There was an error
+        if (test.can_fail) {
+          // Failure is OK
+          expect(hadError).to.equal(true);
+        } else {
+          // Failure is NOT OK
+          throw new Error('We should not have failed but got an error: ' + caughtError.message);
+        }
       }
-      expect(result).to.deep.equal(test.expected);
+
+      result = deepReplaceBuffer(result);
+
+      // in javascript 0 === -0, but in mocha it's not the same.
+      // this normalizes the 0's.
+      if (result === -0) result = 0;
+
+      try {
+        expect(result).to.deep.equal(test.expected);
+      } catch (e) {
+        if (test.can_fail) {
+          // Optional failure
+          this.skip('can_fail was true');
+        } else {
+          throw e;
+        }
+      }
+
     }
   });
 
 }
 
-describe('HTTP-WG tests', () => {
+/**
+ * The HTTP-WG tests decode the "byte sequence" type as a Base32 string.
+ *
+ * We decode them in buffers. This function replaces all Buffers to base32
+ * strings.
+ */
+function deepReplaceBuffer(input) {
 
-  describe('binary.json', () => {
+  if(input instanceof Buffer) {
+    return base32Encode(input, 'RFC4648');
+  }
 
-    for(const test of binaryTests) {
+  if (input === null) {
+    return null;
+  }
 
-      addTest(test);
+  if (typeof input === 'object') {
 
+    for(const [prop, value] of Object.entries(input)) {
+      input[prop] = deepReplaceBuffer(value);
     }
 
-  });
+  }
 
-  describe('list.json', () => {
+  return input;
 
-    for(const test of listTests) {
-
-      addTest(test);
-
-    }
-
-  });
-
-  describe('number.json', () => {
-
-    for(const test of numberTests) {
-
-      addTest(test);
-
-    }
-
-  });
-
-  describe('string.json', () => {
-
-    for(const test of stringTests) {
-
-      addTest(test);
-
-    }
-
-  });
-
-});
+}